--- bin/web2json
@@ -1,5 +1,0 @@
-#!/usr/bin/env ruby
-
-require 'Web2json'
-
-puts Web2json
