+++ lib/higokami/nodeparser.rb
